feat(dump): local working config with many updates
