+++ init.lua
@@ -9,8 +9,7 @@ o = vim.o
 opt = vim.opt
 
 -- Self assigned global variables
--- g.left_sidebar_width = 55
-g.left_sidebar_width = 45
+g.left_sidebar_width = 55
 
 -- Disable providers
 g.loaded_perl_provider = 0
@@ -23,3 +22,5 @@ require 'core.lazy-plugins'
 require 'core.mappings'
 require 'core.neovide'
 require 'core.settings'
+require 'core.lsp'
+-- require 'search_with_index'
