+++ lua/core/mappings.lua
@@ -7,6 +7,10 @@ local map = api.nvim_set_keymap
 keymap.set("n", "<Space>", "<Nop>", { silent = true, remap = false })
 g.mapleader = " "
 
+keymap.set('n', '<leader><leader>', function()
+  cmd('e!')
+end, options)
+
 -- CTRL + Backspace to delete word in nvim
 -- NOTE: Disable Ctrl + Backspace mapping in alacritty
 map('i', "<C-BS>", "<C-W>", options)
@@ -29,8 +33,8 @@ map('n', '0', '^', options)
 map('n', '^', '0', options)
 
 -- Easy buffers
-map('n', '<leader>n', ':bn <cr>', options)
-map('n', '<leader>b', ':bp <cr>', options)
+-- map('n', '<leader>n', ':bn <cr>', options)
+-- map('n', '<leader>b', ':bp <cr>', options)
 
 -- Copy to system clipboard
 if not fn.has('macunix') then
@@ -93,17 +97,13 @@ map('n', '<tab>', '<c-6>', options)
 -- Comment
 -- map('v', "<c-'>", "ysiw'", options)
 
--- LSP
--- vim.keymap.set('n', 'K', vim.lsp.buf.hover, options)
--- vim.keymap.set('n', '<leader>rn', vim.lsp.buf.rename, options)
--- vim.keymap.set('n', '<leader>gr', vim.lsp.buf.references, options)
-
 -- Telescope (File Pickers)
 -- map('n', '<leader>ff', ':lua require("telescope.builtin").find_files() <cr>', options)
 map('n', '<leader>ff', ':lua require("telescope.builtin").find_files({layout_config={height=0.90,width=0.75}}) <cr>', options)
 map('n', '<leader>fg', ':lua require("telescope.builtin").live_grep() <cr>', options)
 map('n', '<leader>fb', ':lua require("telescope.builtin").buffers({layout_config={width=0.5}}) <cr>', options)
 map('n', '<leader>fh', ':lua require("telescope.builtin").help_tags() <cr>', options)
+map('n', '<leader>fl', ':lua require("telescope.builtin").highlights() <cr>', options)
 map('n', '<leader>fo', ':lua require("telescope.builtin").oldfiles() <cr>', options)
 
 -- Telescope (Git)
@@ -128,18 +128,20 @@ map('n', '<leader>fdg', ':GrepInDirectory <cr>', options)
 -- map('n', '<leader>i', ':NvimTreeFindFile <cr>', options)
 
 -- neo-tree
-map('n', '<leader>e', '<cmd>Neotree toggle<cr>', options)
+-- map('n', '<leader>e', '<cmd>Neotree toggle<cr>', options)
+map('n', '<leader>e', '<cmd>DBUIClose<cr><cmd>Neotree toggle<cr>', options)
 map('n', '<leader>i', '<cmd>Neotree reveal<cr>', options)
 
 -- Lazygit
 map('n', '<leader>lg', ':LazyGit <cr>', options)
--- map('n', '<leader>lgc', ':LazyGitFilterCurrentFile <cr>', options)
+map('n', '<leader>lc', ':LazyGitFilterCurrentFile <cr>', options)
+-- map('n', '<leader>lg', '<cmd>lua _lazygit_toggle()<CR>', options)
 
 -- Quick switch to current file scss & js
 -- `:r` removes the file extension
 -- `.scss` or `.js` add files extension to the modified filename
-map('n', '<leader>st', ':e %:r.scss <cr>', options)
-map('n', '<leader>js', ':e %:r.js <cr>', options)
+-- map('n', '<leader>st', ':e %:r.scss <cr>', options)
+-- map('n', '<leader>js', ':e %:r.js <cr>', options)
 
 -- Quick switch to current file scss & js
 -- `:e` is edit command of vim
@@ -152,7 +154,8 @@ map('n', '<leader>js', ':e %:r.js <cr>', options)
 -- Ruby on Rails: quick switch to rspec
 -- `%:s/^app/spec/` Replaces the app/ directory with spec/.
 -- `:r_spec.rb` Removes the file extension (.rb) and adds _spec.rb to the filename.
-map('n', '<leader>test', ':e %:s/^app/spec/:r_spec.rb <cr>', options)
+-- map('n', '<leader>test', ':e %:s/^app/spec/:r_spec.rb <cr>', options)
+map('n', '<leader>test', ':e %:s/app/spec/:r_spec.rb <cr>', options)
 
 -- Dadbod UI
 -- map('n', '<leader>d', '<cmd>NvimTreeClose<cr><cmd>DBUIToggle<cr>', options)
@@ -169,51 +172,53 @@ map('n', '<leader>z', '<cmd>ZenMode<cr>', options)
 -- Undotree
 map('n', '<leader>u', ':lua require("undotree").toggle() <cr>', options)
 -- ToggleTerm
--- map('n', '<leader>t', ':ToggleTerm<cr>', options)
+map('n', '<leader>ft', ':ToggleTerm<cr>', options)
 
--- Debugging
+-- START Debugging
+-- Chainsaw
 map('n', '<leader>cl', '<cmd>lua require("chainsaw").variableLog()<cr>', options)
 map('n', '<leader>cc', '<cmd>lua require("chainsaw").clearLog()<cr>', options)
 map('n', '<leader>ce', '<cmd>lua require("chainsaw").emojiLog()<cr>', options)
 map('n', '<leader>cr', '<cmd>lua require("chainsaw").removeLogs()<cr>', options)
 
--- LSP Diagnostic
-keymap.set('n', '<leader>j', function()
-  vim.diagnostic.goto_next()
+-- vim-test (list of commands: https://github.com/vim-test/vim-test?tab=readme-ov-file#setup)
+map('n', '<leader>tt', '<cmd>TestNearest<cr>', options)
+map('n', '<leader>TT', '<cmd>TestFile<cr>', options)
+
+-- END Debugging
+
+-- START: MultiCursor
+local mc = require("multicursor-nvim")
+
+-- Add or skip adding a new cursor by matching word/selection
+keymap.set({'n', 'v'}, '<leader>n', function()
+  mc.matchAddCursor(1)
 end, options)
 
--- keymap.set('n', '<leader>k', function()
---   vim.diagnostic.goto_prev()
--- end, options)
+-- END: MultiCursor
 
-keymap.set('n', '<leader>k', function()
-  vim.diagnostic.open_float()
-end, options)
+-- Bufferline
+map('n', '<leader>1', '<cmd>BufferLineGoToBuffer 1<cr>', options)
+map('n', '<leader>2', '<cmd>BufferLineGoToBuffer 2<cr>', options)
+map('n', '<leader>3', '<cmd>BufferLineGoToBuffer 3<cr>', options)
+map('n', '<leader>4', '<cmd>BufferLineGoToBuffer 4<cr>', options)
+map('n', '<leader>5', '<cmd>BufferLineGoToBuffer 5<cr>', options)
+map('n', '<leader>6', '<cmd>BufferLineGoToBuffer 6<cr>', options)
+map('n', '<leader>7', '<cmd>BufferLineGoToBuffer 7<cr>', options)
+map('n', '<leader>8', '<cmd>BufferLineGoToBuffer 8<cr>', options)
+map('n', '<leader>9', '<cmd>BufferLineGoToBuffer 9<cr>', options)
+map('n', '<leader>$', '<cmd>BufferLineGoToBuffer $<cr>', options)
 
--- LSP things
-map('n', '<c-]', '<cmd>lua vim.lsp.buf.definition()<cr>', options)
--- map('n', '<leader>gd', '<cmd>lua vim.lsp.buf.definition()<cr>', options)
-map('n', '<leader>ft', '<cmd>lua vim.lsp.buf.format()<cr>', options)
-map('n', '<leader>ca', '<cmd>lua vim.lsp.buf.code_action()<cr>', options)
-map('i', '<c-s>', '<cmd>lua vim.lsp.buf.signature_help()<cr>', options)
-keymap.set('n', '<c-t>', function()
-  local filetype = vim.bo.filetype
+-- QuickNote (mapping inspiration got from vim-bookmark, just changed the `m` to `s`)
+map('n', '<leader>st', '<cmd>lua require("quicknote").ToggleNoteSigns()<cr>', options)
+map('n', '<leader>ss', '<cmd>lua require("quicknote").NewNoteAtCurrentLine()<cr>', options)
+map('n', '<leader>sd', '<cmd>lua require("quicknote").DeleteNoteAtCurrentLine()<cr>', options)
+map('n', '<leader>sf', '<cmd>Telescope quicknote scope=CurrentBuffer<cr>', options)
+map('n', '<leader>sa', '<cmd>lua require("quicknote").ListNotesForCurrentBuffer()<cr>', options)
+map('n', '<leader>sn', '<cmd>lua require("quicknote").JumpToNextNote()<cr>', options)
+-- map('n', '<leader>sp', '<cmd>lua require("quicknote").JumpToPreviousNote()<cr>', options)
 
-  cmd('vsplit')
-  if filetype == 'markdown' then
-    cmd('ObsidianFollowLink')
-  else
-    -- Use api.nvim_feedkeys to simulate pressing <c-]>
-    api.nvim_feedkeys(api.nvim_replace_termcodes("<c-]>", true, true, true), 'n', true)
-  end
-end, options)
-
--- LSP TSX
-map('n', '<leader>tsi', '<cmd>TSToolsAddMissingImports<cr>', options)
-map('n', '<leader>tss', '<cmd>TSToolsSortImports<cr>', options)
-map('n', '<leader>tsr', '<cmd>TSToolsRemoveUnusedImports<cr>', options)
-map('n', '<leader>tsl', '<cmd>TSToolsFileReferences<cr>', options)
-
+-- TODO: DRAFT
 -- keymap.set('n', '<c-a>', vim.lsp.buf.code_action, bufopts) -- this is working (but c-a mapping is for increment number)
 -- g.mapping_code_action = { mode = 'n', mapping = '<c-a>', command = vim.lsp.buf.code_action }
 -- g.mapping_buf_hover = keymap.set('n', 'K', vim.lsp.buf.hover, bufopts)
