+++ lua/plugins/archive/debugging/neotest.lua
@@ -1,61 +1,67 @@
-return {
-  -- {
-  --   "nvim-neotest/neotest",
-  --   lazy = true,
-  --   dependencies = {
-  --     -- Default dependencies
-  --     "nvim-neotest/nvim-nio",
-  --     "nvim-lua/plenary.nvim",
-  --     "antoinemadec/FixCursorHold.nvim",
-  --     "nvim-treesitter/nvim-treesitter",
-  --
-  --     -- Runners
-  --     "olimorris/neotest-rspec",
-  --   },
-  --   config = function ()
-  --     local neotest = require("neotest")
-  --
-  --     neotest.setup({
-  --       adapters = {
-  --         output_panel = true,
-  --         require("neotest-rspec")({
-  --           rspec_cmd = function ()
-  --             -- return vim.tbl_flatten({
-  --             --   "bundle",
-  --             --   "exec",
-  --             --   "rspec",
-  --             -- })
-  --             return vim.tbl_flatten({
-  --               "docker",
-  --               "compose",
-  --               "exec",
-  --               "-i",
-  --               "-w", "/app",
-  --               "-e", "RAILS_ENV=test",
-  --               "merchants_api",
-  --               "bundle",
-  --               "exec",
-  --               "rspec"
-  --             })
-  --           end,
-  --           filter_dirs = { ".git", "node_modules" },
-  --
-  --           -- transform_spec_path = function (path)
-  --           --   return path
-  --           -- end,
-  --           -- results_path = function ()
-  --           --   return async.fn.tempname()
-  --           -- end,
-  --
-  --           transform_spec_path = function(path)
-  --             local prefix = require('neotest-rspec').root(path)
-  --             return string.sub(path, string.len(prefix) + 2, -1)
-  --           end,
-  --
-  --           results_path = "tmp/rspec.output"
-  --         }),
-  --       },
-  --     })
-  --   end
-  -- }
+-- TODO: add mapping for the test commands
+-- commands list: https://github.com/nvim-neotest/neotest?tab=readme-ov-file#usage
+
+local M = {
+  "nvim-neotest/neotest",
+  lazy = true,
+  dependencies = {
+    -- Default dependencies
+    "nvim-neotest/nvim-nio",
+    "nvim-lua/plenary.nvim",
+    "antoinemadec/FixCursorHold.nvim",
+    "nvim-treesitter/nvim-treesitter",
+
+    -- Runners
+    "olimorris/neotest-rspec",
+  }
 }
+
+function M.config()
+  local neotest = require("neotest")
+
+  neotest.setup({
+    adapters = {
+      output_panel = true,
+      require("neotest-rspec")({
+        rspec_cmd = function ()
+          return vim.tbl_flatten({
+            "bundle",
+            "exec",
+            "rspec",
+          })
+
+          -- -- NOTE: For Docker
+          -- return vim.tbl_flatten({
+          --   "docker",
+          --   "compose",
+          --   "exec",
+          --   "-i",
+          --   "-w", "/app",
+          --   "-e", "RAILS_ENV=test",
+          --   "merchants_api",
+          --   "bundle",
+          --   "exec",
+          --   "rspec"
+          -- })
+        end,
+        filter_dirs = { ".git", "node_modules" },
+
+        -- transform_spec_path = function (path)
+        --   return path
+        -- end,
+        -- results_path = function ()
+        --   return async.fn.tempname()
+        -- end,
+
+        transform_spec_path = function(path)
+          local prefix = require('neotest-rspec').root(path)
+          return string.sub(path, string.len(prefix) + 2, -1)
+        end,
+
+        results_path = "tmp/rspec.output"
+      }),
+    },
+  })
+end
+
+return M
