--- lua/plugins/archive/toggleterm.lua
@@ -1,30 +1,0 @@
-return {
-  -- {
-  --   'akinsho/toggleterm.nvim',
-  --   version = "*",
-  --   config = function()
-  --     local tt = require 'toggleterm'
-  --
-  --     tt.setup({
-  --       direction = 'float',
-  --       float_opts = {
-  --         border = 'curved',
-  --         winblend = 3,
-  --         title_pos = 'center'
-  --       },
-  --     })
-  --
-  --     function _G.set_terminal_keymaps()
-  --       local opts = { noremap = true }
-  --       vim.api.nvim_buf_set_keymap(0, 't', '<esc>', [[<C-\><C-n>]], opts)
-  --       vim.api.nvim_buf_set_keymap(0, "t", "jk", [[<C-\><C-n>]], opts)
-  --       vim.api.nvim_buf_set_keymap(0, "t", "<C-h>", [[<C-\><C-n><C-W>h]], opts)
-  --       vim.api.nvim_buf_set_keymap(0, "t", "<C-j>", [[<C-\><C-n><C-W>j]], opts)
-  --       vim.api.nvim_buf_set_keymap(0, "t", "<C-k>", [[<C-\><C-n><C-W>k]], opts)
-  --       vim.api.nvim_buf_set_keymap(0, "t", "<C-l>", [[<C-\><C-n><C-W>l]], opts)
-  --     end
-  --
-  --     vim.cmd "autocmd! TermOpen term://* lua set_terminal_keymaps()"
-  --   end
-  -- }
-}
