+++ lua/plugins/colorschemes/init.lua
@@ -1,7 +1,9 @@
 return {
-  -- require 'plugins.colorschemes.catppuccin',
+  require 'plugins.colorschemes.catppuccin',
+  -- require 'plugins.colorschemes.catppuccin-mod',
   -- require 'plugins.colorschemes.deepwhite',
-  require 'plugins.colorschemes.nightfox',
+  -- require 'plugins.colorschemes.nightfox',
+  -- require 'plugins.colorschemes.vividchalk',
   -- require 'plugins.colorschemes.papercolor',
   -- require 'plugins.colorschemes.tokyo-night'
 }
