+++ lua/plugins/colorschemes/vividchalk.lua
@@ -1,0 +1,10 @@
+return {
+  {
+    "tpope/vim-vividchalk",
+    lazy = false,
+    priority = 1000,
+    config = function()
+      vim.cmd.colorscheme("vividchalk")
+    end
+  }
+}
