+++ lua/plugins/debugging/chainsaw.lua
@@ -4,13 +4,31 @@ local M = {
 
 function M.config()
   require("chainsaw").setup {
-    marker = "🪚",
+    -- marker = "🪚",
+    marker = "︻デ┳═ー",
+
     logHighlightGroup = false,
     logStatements = {
       variableLog = {
         javascript = "console.log('%s %s:', %s);",
+        ruby = {
+          "puts '=========================================='",
+          "pp '{{marker}} {{var}}:', {{var}}",
+          "puts '=========================================='",
+        },
       },
-    }
+      emojiLog = {
+        ruby = {
+          "puts '{{marker}}: {{emoji}}'"
+        },
+      },
+    },
+    -- configuration for specific logtypes
+    logTypes = {
+      emojiLog = {
+        emojis = { "ʘ‿ʘ", "ಥ‿ಥ", "(ツ)", "(⌐■_■)", "(－_－)", "ಠ_ಠ" },
+      },
+    },
   }
 end
