+++ lua/plugins/image.lua
@@ -48,7 +48,8 @@ function M.config()
     max_width_window_percentage = nil,
     max_height_window_percentage = 50,
     window_overlap_clear_enabled = false,                                               -- toggles images when windows are overlapped
-    window_overlap_clear_ft_ignore = { "cmp_menu", "cmp_docs", "snacks_notif", "scrollview", "scrollview_sign" },
+    -- window_overlap_clear_ft_ignore = { "cmp_menu", "cmp_docs", "snacks_notif", "scrollview", "scrollview_sign" },
+    window_overlap_clear_ft_ignore = { "snacks_notif", "scrollview", "scrollview_sign" },
     editor_only_render_when_focused = false,                                            -- auto show/hide images when the editor gains/looses focus
     tmux_show_only_in_active_window = false,                                            -- auto show/hide images in the correct Tmux window (needs visual-activity off)
     hijack_file_patterns = { "*.png", "*.jpg", "*.jpeg", "*.gif", "*.webp", "*.avif" }, -- render image files as images when opened
