+++ lua/plugins/nice-to-have/harpoon.lua
@@ -2,6 +2,7 @@ local M = {
   "ThePrimeagen/harpoon",
   branch = "harpoon2",
   dependencies = { "nvim-lua/plenary.nvim" },
+  enabled = false
 }
 
 function M.config()
