+++ lua/plugins/toggleterm.lua
@@ -1,0 +1,61 @@
+local M = {
+  "akinsho/toggleterm.nvim",
+  version = "*",
+  enabled = false
+}
+
+function M.config()
+  local tt = require("toggleterm")
+
+  tt.setup {
+    -- direction = "float",
+    -- float_opts = {
+    --   border = "curved",
+    --   winblend = 3,
+    --   title_pos = "center"
+    -- },
+    open_mapping = [[<c-\>]],
+    insert_mappings = true,
+    terminal_mappings = true,
+  }
+
+  function _G.set_terminal_keymaps()
+    local opts = { noremap = true }
+    vim.api.nvim_buf_set_keymap(0, "t", "<esc>", [[<C-\><C-n>]], opts)
+    vim.api.nvim_buf_set_keymap(0, "t", "jk", [[<C-\><C-n>]], opts)
+    vim.api.nvim_buf_set_keymap(0, "t", "<C-h>", [[<C-\><C-n><C-W>h]], opts)
+    vim.api.nvim_buf_set_keymap(0, "t", "<C-j>", [[<C-\><C-n><C-W>j]], opts)
+    vim.api.nvim_buf_set_keymap(0, "t", "<C-k>", [[<C-\><C-n><C-W>k]], opts)
+    -- disable "<C-l>" to enabled "clear" inside terminal
+    -- vim.api.nvim_buf_set_keymap(0, "t", "<C-l>", [[<C-\><C-n><C-W>l]], opts)
+  end
+
+  vim.cmd "autocmd! TermOpen term://* lua set_terminal_keymaps()"
+
+  -- Lazygit
+  local Terminal  = require('toggleterm.terminal').Terminal
+
+  local lazygit = Terminal:new({
+    cmd = "lazygit",
+    dir = "git_dir",
+    direction = "float",
+    float_opts = {
+      border = "single",
+    },
+    -- function to run on opening the terminal
+    on_open = function(term)
+      vim.cmd("startinsert!")
+      vim.api.nvim_buf_set_keymap(term.bufnr, "n", "q", "<cmd>close<CR>", {noremap = true, silent = true})
+    end,
+    -- function to run on closing the terminal
+    on_close = function(term)
+      vim.cmd("startinsert!")
+    end,
+  })
+
+  function _lazygit_toggle()
+    lazygit:toggle()
+  end
+end
+
+return M
